REUPLOAD of project with new scope
Due to time constraints, we had to scale our app back a little, from showing graphs of datapoints to due with the stock to just a snapshot of the latest price. The search function and all other facts of the app work, with finalProject2 being the final project and finalProject1 being the initial idea that will be worked on eventually.